工作流库0216/T44-海外债/requirements.txt
# 海外债项目依赖列表
# Overseas Bond Project Dependencies

# 数据处理
pandas>=1.5.0
numpy>=1.21.0
chardet>=5.0.0

# 数据库
pymysql>=1.0.0
sqlalchemy>=1.4.0

# Excel处理
openpyxl>=3.0.0
xlrd>=2.0.0

# 环境变量管理
python-dotenv>=0.20.0

# 可视化
matplotlib>=3.5.0

# 可选依赖（根据实际数据源启用）
# Wind金融终端
# WindPy>=1.0.0

# 同花顺iFinD
# iFinDPy>=1.0.0

# 网络请求（如需要）
requests>=2.28.0

# 开发工具
ipykernel>=6.0.0
jupyter>=1.0.0
